# Conflicts:
#	ProyectoLobo/Assets/Resources/FogOfWarIV.prefab
#	ProyectoLobo/Assets/Scenes/EscenarioV1.0.unity

+++ ProyectoLobo/Assets/Scripts/Visibility/VisibilityCone.cs
@@ -13,7 +13,7 @@ public class VisibilityCone:MonoBehaviour  {
     public static List<Vector2> pixelsColored;
     [Range(1,4)]
     public int sampleModifier=2;
-    public int Radius = 200;
+    public float Radius = 200;
 
     public Color ColorOfCone;
     public Color ColorOfPath;
